Django 1.7 support, fix for Django Ticket #21057
Django's FileField used to leave temporary files around. ImageKit made use of this by not re-opening the file. This bug was fixed in https://code.djangoproject.com/ticket/21057 though, so content.seek(0) will fail (content is unlinked during self.storage.save()).

I'm not sure if this is the best way to do this, but just want to draw your attention to this issue.

Here is the commit that broke it:
https://github.com/django/django/commit/30fc49a7ca0d030c7855f31ed44395903fa6abdd